Extract logic to add behavior to object in utils file
