+++ newIDE/app/src/BehaviorsEditor/index.js
@@ -9,7 +9,6 @@ import IconButton from '../UI/IconButton';
 import EmptyMessage from '../UI/EmptyMessage';
 import { MiniToolbarText } from '../UI/MiniToolbar';
 import HelpIcon from '../UI/HelpIcon';
-import newNameGenerator from '../Utils/NewNameGenerator';
 import NewBehaviorDialog from './NewBehaviorDialog';
 import BehaviorsEditorService from './BehaviorsEditorService';
 import Window from '../Utils/Window';
@@ -35,6 +34,7 @@ import {
 } from '../Utils/GDevelopServices/Badge';
 import AuthenticatedUserContext from '../Profile/AuthenticatedUserContext';
 import { getBehaviorTutorialIds } from '../Utils/GDevelopServices/Tutorial';
+import { addBehaviorToObject } from '../Utils/Behavior';
 
 const gd: libGDevelop = global.gd;
 
@@ -60,33 +60,10 @@ const BehaviorsEditor = (props: Props) => {
 
   const { values } = React.useContext(PreferencesContext);
 
-  const hasBehaviorWithType = (type: string) => {
-    return allBehaviorNames
-      .map(behaviorName => object.getBehavior(behaviorName))
-      .map(behavior => behavior.getTypeName())
-      .filter(behaviorType => behaviorType === type).length;
-  };
-
   const _addBehavior = (type: string, defaultName: string) => {
     setNewBehaviorDialogOpen(false);
 
-    if (hasBehaviorWithType(type)) {
-      const answer = Window.showConfirmDialog(
-        "There is already a behavior of this type attached to the object. It's possible to add this behavior again, but it's unusual and may not be always supported properly. Are you sure you want to add this behavior again?"
-      );
-
-      if (!answer) return;
-    }
-
-    const name = newNameGenerator(defaultName, name =>
-      object.hasBehaviorNamed(name)
-    );
-    gd.WholeProjectRefactorer.addBehaviorAndRequiredBehaviors(
-      project,
-      object,
-      type,
-      name
-    );
+    addBehaviorToObject(project, object, type, defaultName);
 
     forceUpdate();
     if (props.onSizeUpdated) props.onSizeUpdated();
